+++ layers/Schema/packages/media/src/FieldResolvers/ObjectType/MediaObjectTypeFieldResolver.php
@@ -96,9 +96,9 @@ class MediaObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResol
             'altText',
             'description',
             'date',
-            'dateAsString',
-            'modified',
-            'modifiedAsString',
+            'dateStr',
+            'modifiedDate',
+            'modifiedDateStr',
             'mimeType',
         ];
     }
@@ -116,9 +116,9 @@ class MediaObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResol
             'altText' => $this->getStringScalarTypeResolver(),
             'description' => $this->getStringScalarTypeResolver(),
             'date' => $this->getDateTimeScalarTypeResolver(),
-            'dateAsString' => $this->getStringScalarTypeResolver(),
-            'modified' => $this->getDateTimeScalarTypeResolver(),
-            'modifiedAsString' => $this->getStringScalarTypeResolver(),
+            'dateStr' => $this->getStringScalarTypeResolver(),
+            'modifiedDate' => $this->getDateTimeScalarTypeResolver(),
+            'modifiedDateStr' => $this->getStringScalarTypeResolver(),
             'mimeType' => $this->getStringScalarTypeResolver(),
             default => parent::getFieldTypeResolver($objectTypeResolver, $fieldName),
         };
@@ -129,9 +129,9 @@ class MediaObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResol
         return match ($fieldName) {
             'src',
             'date',
-            'dateAsString',
-            'modified',
-            'modifiedAsString'
+            'dateStr',
+            'modifiedDate',
+            'modifiedDateStr'
                 => SchemaTypeModifiers::NON_NULLABLE,
             default
                 => parent::getFieldTypeModifiers($objectTypeResolver, $fieldName),
@@ -151,9 +151,9 @@ class MediaObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResol
             'altText' => $this->getTranslationAPI()->__('Media element alt text', 'pop-media'),
             'description' => $this->getTranslationAPI()->__('Media element description', 'pop-media'),
             'date' => $this->getTranslationAPI()->__('Media element\'s published date', 'pop-media'),
-            'dateAsString' => $this->getTranslationAPI()->__('Media element\'s published date, in String format', 'pop-media'),
-            'modified' => $this->getTranslationAPI()->__('Media element\'s modified date', 'pop-media'),
-            'modifiedAsString' => $this->getTranslationAPI()->__('Media element\'s modified date, in String format', 'pop-media'),
+            'dateStr' => $this->getTranslationAPI()->__('Media element\'s published date, in String format', 'pop-media'),
+            'modifiedDate' => $this->getTranslationAPI()->__('Media element\'s modified date', 'pop-media'),
+            'modifiedDateStr' => $this->getTranslationAPI()->__('Media element\'s modified date, in String format', 'pop-media'),
             'mimeType' => $this->getTranslationAPI()->__('Media element\'s mime type', 'pop-media'),
             default => parent::getFieldDescription($objectTypeResolver, $fieldName),
         };
@@ -187,9 +187,9 @@ class MediaObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResol
     {
         return match ($fieldName) {
             'date' => [CommonFilterInputContainerModuleProcessor::class, CommonFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_GMTDATE],
-            'dateAsString' => [CommonFilterInputContainerModuleProcessor::class, CommonFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_GMTDATE_AS_STRING],
-            'modified' => [CommonFilterInputContainerModuleProcessor::class, CommonFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_GMTDATE],
-            'modifiedAsString' => [CommonFilterInputContainerModuleProcessor::class, CommonFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_GMTDATE_AS_STRING],
+            'dateStr' => [CommonFilterInputContainerModuleProcessor::class, CommonFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_GMTDATE_AS_STRING],
+            'modifiedDate' => [CommonFilterInputContainerModuleProcessor::class, CommonFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_GMTDATE],
+            'modifiedDateStr' => [CommonFilterInputContainerModuleProcessor::class, CommonFilterInputContainerModuleProcessor::MODULE_FILTERINPUTCONTAINER_GMTDATE_AS_STRING],
             default => parent::getFieldFilterInputContainerModule($objectTypeResolver, $fieldName),
         };
     }
@@ -238,14 +238,14 @@ class MediaObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResol
                 return $this->getMediaTypeAPI()->getDescription($media);
             case 'date':
                 return new DateTime($this->getMediaTypeAPI()->getDate($media, $fieldArgs['gmt']));
-            case 'dateAsString':
+            case 'dateStr':
                 return $this->getDateFormatter()->format(
                     $fieldArgs['format'],
                     $this->getMediaTypeAPI()->getDate($media, $fieldArgs['gmt'])
                 );
-            case 'modified':
+            case 'modifiedDate':
                 return new DateTime($this->getMediaTypeAPI()->getModified($media, $fieldArgs['gmt']));
-            case 'modifiedAsString':
+            case 'modifiedDateStr':
                 return $this->getDateFormatter()->format(
                     $fieldArgs['format'],
                     $this->getMediaTypeAPI()->getModified($media, $fieldArgs['gmt'])